requirements.txt
# Основные зависимости
python-dotenv==1.0.0
openai==1.12.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.2.0
openpyxl==3.1.2

# Google Sheets API
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0

# Планировщик
schedule==1.2.1

# Логирование
loguru==0.7.2

# Утилиты
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2024.1 